+++ src/net4cxx/plugins/web/httpserver.cpp
@@ -230,7 +230,7 @@ void HTTPConnection::startRequest() {
 
 std::string HTTPConnection::formatChunk(const Byte *data, size_t length) {
     if (_expectedContentRemaining) {
-        _expectedContentRemaining = *_expectedContentRemaining - length;
+        _expectedContentRemaining = *_expectedContentRemaining - (ssize_t)length;
         if (*_expectedContentRemaining < 0) {
             try {
                 NET4CXX_THROW_EXCEPTION(HTTPOutputError, "Tried to write more data than Content-Length");
